Move reporting of modules
